Move translated constants to new translations file
